refactor: move types to `types` directory
