+++ src/core/mixins/MediaInfo.ts
@@ -1,7 +1,7 @@
 import type { ApiResponse } from '../Actions.js';
 import type Actions from '../Actions.js';
 import * as Constants from '../../utils/Constants.js';
-import type { DownloadOptions, FormatFilter, FormatOptions, URLTransformer } from '../../utils/FormatUtils.js';
+import type { DownloadOptions, FormatFilter, FormatOptions, URLTransformer } from '../../types/FormatUtils.js';
 import * as FormatUtils from '../../utils/FormatUtils.js';
 import { InnertubeError } from '../../utils/Utils.js';
 import type Format from '../../parser/classes/misc/Format.js';
