+++ src/types/index.ts
@@ -4,3 +4,4 @@ export * from './Cache.js';
 export * from './PlatformShim.js';
 export * from './Clients.js';
 export * from './Endpoints.js';
+export * from './FormatUtils.js';
